automatically calls script main now
